+++ app/templates/client/scripts/backbone/browserify/main/main.js
@@ -57,8 +57,10 @@ user.isAuthenticated({
         var pushState = !!(enablePushState && window.history && window.history.pushState);
 
         if (pushState) {
+            // Start listening to route changes with pushState
             Backbone.history.start({ pushState: true, root: '/' });
         } else {
+            // Start listening to route changes without pushState
             Backbone.history.start();
         }
 
@@ -69,7 +71,7 @@ user.isAuthenticated({
 
     }
 
-});<% } else { %>
+});<% } else if (useServer) { %>
 // Enable pushState for compatible browsers
 var enablePushState = true;
 
@@ -77,15 +79,20 @@ var enablePushState = true;
 var pushState = !!(enablePushState && window.history && window.history.pushState);
 
 if (pushState) {
+    // Start listening to route changes with pushState
     Backbone.history.start({ pushState: true, root: '/' });
 } else {
+    // Start listening to route changes without pushState
     Backbone.history.start();
 }
 
 // Handle pushState for incompatible browsers (IE9 and below)
 if (!pushState && window.location.pathname !== '/') {
     window.location.replace('/#' + window.location.pathname);
-}<% } %>
+}<% } else { %>
+
+// Start listening to route changes
+Backbone.history.start();<% } %>
 
 // Set up global click event handler to use pushState for links
 // use 'data-bypass' attribute on anchors to allow normal link behavior