+++ app/templates/client/scripts/backbone/requirejs/main/main.js
@@ -76,7 +76,7 @@ define('init', function(require) {
 
         }
 
-    });<% } else { %>
+    });<% } else if (useServer) { %>
     // Enable pushState for compatible browsers
     var enablePushState = true;
 
@@ -84,15 +84,20 @@ define('init', function(require) {
     var pushState = !!(enablePushState && window.history && window.history.pushState);
 
     if (pushState) {
+        // Start listening to route changes with pushState
         Backbone.history.start({ pushState: true, root: '/' });
     } else {
+        // Start listening to route changes without pushState
         Backbone.history.start();
     }
 
     // Handle pushState for incompatible browsers (IE9 and below)
     if (!pushState && window.location.pathname !== '/') {
         window.location.replace('/#' + window.location.pathname);
-    }<% } %>
+    }<% } else { %>
+
+    // Start listening to route changes
+    Backbone.history.start();<% } %>
 
     // Set up global click event handler to use pushState for links
     // use 'data-bypass' attribute on anchors to allow normal link behavior