+++ app/templates/client/scripts/react/actions/messages.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var Dispatcher = require('../dispatchers/default');
-var ActionTypes = require('../constants/action-types');
+var messagesConstants = require('../constants/messages');
 var messagesDefaults = require('../constants/defaults').messages;
 var assign = require('object-assign');
 
@@ -9,7 +9,7 @@ module.exports = {
 
     setMessages: function(messages) {
         Dispatcher.handleViewAction({
-            actionType: ActionTypes.SET_MESSAGES,
+            actionType: messagesConstants.SET_MESSAGES,
             messages: assign({}, messagesDefaults, messages)
         });
     }