+++ app/templates/client/scripts/react/constants/messages.js
@@ -1,0 +1,12 @@
+'use strict';
+
+var keyMirror = require('keymirror');
+
+var messagesConstants = keyMirror({
+
+    // Global message types
+    SET_MESSAGES: null
+
+});
+
+module.exports = messagesConstants;