+++ app/templates/client/scripts/react/constants/page.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var keyMirror = require('keymirror');
+
+var pageConstants = keyMirror({
+
+    // Page action types
+    SET_CURRENT_PAGE: null
+
+});
+
+module.exports = pageConstants;
+