+++ app/templates/client/scripts/react/constants/routes.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var keyMirror = require('keymirror');
+
+var routeConstants = keyMirror({
+
+    // Route action types
+    SET_CURRENT_ROUTE: null
+
+});
+
+module.exports = routeConstants;
+