+++ app/templates/client/scripts/react/constants/user.js
@@ -1,0 +1,12 @@
+'use strict';
+
+var keyMirror = require('keymirror');
+
+var userConstants = keyMirror({
+
+    // User action types
+    SET_CURRENT_USER: null,
+
+});
+
+module.exports = userConstants;